requirements.txt
pandas
streamlit
spacy
transformers
sentencepiece
googletrans==3.1.0a0
pyvespa
